moves message fixture to insecure package
